+++ ground/openpilotgcs/src/libs/opmapcontrol/src/internals/projections/platecarreeprojection.cpp
@@ -82,7 +82,7 @@ double PlateCarreeProjection::Flattening() const
 }
 Size PlateCarreeProjection::GetTileMatrixMaxXY(const int &zoom)
 {
-    int y = (int)pow(2, zoom);
+    int y = (int)pow((float)2, zoom);
 
     return Size((2 * y) - 1, y - 1);
 }
